+++ benchmark/README.md
@@ -1,0 +1,1 @@
+Usage: python benchmark/run.py structured_data_classification report.csv
